用户注册功能

+++ src/main/java/com/miaoshaproject/service/model/UserModel.java
@@ -44,7 +44,7 @@ public class UserModel {
         return gender;
     }
 
-    public void setGender(Integer gender) {
+    public void setGender(Byte gender) {
         this.gender = gender;
     }
